packages/sample/handler/requirements.txt
gspread==5.10.0
requests==2.32.3
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1  # Keep if you plan OAuth flows, optional otherwise
python-dotenv==1.0.1